requirements.txt
aiohttp==3.6.2
aiohttp-jinja2==1.2.0
aiohttp-session==2.9.0
aiomysql==0.0.20
alembic==1.3.1
async-timeout==3.0.1
attrs==19.3.0
bleach==3.1.0
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
dotenv-python==0.0.1
humanize==0.5.1
idna==2.8
Jinja2==2.10.3
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
multidict==4.6.1
mysql-connector-python==8.0.18
mysql-connector-python-rf==2.2.2
protobuf==3.11.1
pycparser==2.19
PyMySQL==0.9.2
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
six==1.13.0
SQLAlchemy==1.3.11
webencodings==0.5.1
WTForms==2.2.1
yarl==1.4.1
